move and rename mixin file
